Move plugin loader api out into a separate class.
This also makes PluginManager much easier to test.
