+++ chunky/src/java/se/llbit/chunky/plugin/loader/PluginLoader.java
@@ -4,19 +4,15 @@ import se.llbit.chunky.Plugin;
 import se.llbit.chunky.plugin.PluginApi;
 import se.llbit.chunky.plugin.manifest.PluginManifest;
 
-import java.util.Set;
 import java.util.function.BiConsumer;
 
 @PluginApi
 public interface PluginLoader {
   /**
-   * Loads a set of plugins given their manifests.
-   * On successfully loading each plugin class, the plugin instance and its manifest are sent to the onLoad consumer.
-   *
-   * @param pluginManifests the plugins to load.
-   * @param onLoad a function object that the plugin instance and manifest
-   * are sent to after it successfully loads.
+   * Load the plugin specified in the manifest
+   * @param onLoad The consumer to call with the loaded plugin
+   * @param pluginManifest The plugin to load.
    */
   @PluginApi
-  void load(Set<PluginManifest> pluginManifests, BiConsumer<Plugin, PluginManifest> onLoad);
+  void load(BiConsumer<Plugin, PluginManifest> onLoad, PluginManifest pluginManifest);
 }
